Fix a logic bug in the content negotiation code
This caused a bunch of errors on $WORK's web service: Can't use string ("1")
as an ARRAY ref while "strict refs" in use at
/opt/perl5.16.3/lib/site_perl/5.16.3/HTTP/Headers/ActionPack/ContentNegotiation.pm
line 160.

I cannot reproduce this with a test, though. It might also depend on the
version of List::Util being used. Maybe different versions leave $_ in a
different state?